requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
<<<<<<< HEAD
yt-dlp==2024.07.01
pytgcalls==2.2.5
aiohttp==3.9.5
python-dotenv==1.0.1
pymongo
ffmpeg-python
=======
git+https://github.com/pytgcalls/pytgcalls@master
python-dotenv==1.0.1
motor==3.4.0
aiohttp==3.9.3
youtube_dl==2021.12.17
yt-dlp==2024.05.27
pymongo[srv]==4.6.3
>>>>>>> 6f44fcf168a5bb46336301c0e48b584e74efcfc4
